requirements.txt
lumibot
python-dotenv
yfinance
alpaca-py
joblib
ta
scikit-learn
transformers
torch
pandas
numpy

